feat(options): override embedding options via url query
You can now override the plug-in wide embedding options on a per-sandbox basis by specifying them in the link url query:

```js
// gatsby-config.js
{
  resolve: ‘gatsby-remark-embedded-codesandbox’,
  options: {
    embedOptions: { view: 'preview', }
  }
}
```

These embedOptions are applied to every sandbox, but you may want to override them in a single sandbox.
To do this simply add a url query in the generating link:

```md
[code example](embedded-codesandbox://some/example?view=split)
```

The `urlQuery` options will be merged with the `gatsby-config` one.