feat(*): look for manifest file in parent folders
Add simple logic to resolve the `package.json` file in parent folders when not found in the example one.

This lets you define a single (or not) shared file for multiple examples. It will look for the file in parent folder and work its way up until it reaches the **examples root folder**, and if nothing was found it will fallback to the bare minimum allowed:

```json
{
  “name”: “example”
}
```

This also ensures that every sandbox works regardless.

Closes #1